Major refactor of artifacts to minimize runtime dependencies.
This change addresses issue #1. The project is now split into three
subprojects.

  `runtime`: Contains classes that are common to both the api module and
  the generator module. Most of the conversion between ModelNode and
  Java entities is handled by this module.

  `generator`: Creates a `Generator` class that can connect to a running
  WildFly instance to generate Java API classes.

  `api`: The Java configuration API that is generated using the
  `Generator` class.
