+++ api/src/test/java/org/wildfly/swarm/config/IntegrationTestCase.java
@@ -7,8 +7,8 @@ import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
 import org.wildfly.apigen.generator.Config;
-import org.wildfly.apigen.invocation.ClientFactory;
-import org.wildfly.apigen.invocation.EntityAdapter;
+import org.wildfly.apigen.generator.ClientFactory;
+import org.wildfly.config.invocation.EntityAdapter;
 import org.wildfly.swarm.config.datasources.subsystem.dataSource.DataSource;
 import org.wildfly.swarm.config.datasources.subsystem.dataSource.connectionProperties.ConnectionProperties;
 import org.wildfly.swarm.config.logging.subsystem.rootLogger.Root;
