+++ generator/src/main/java/org/wildfly/apigen/generator/Generator.java
@@ -4,8 +4,7 @@ import org.jboss.as.controller.client.ModelControllerClient;
 import org.jboss.dmr.ModelNode;
 import org.jboss.forge.roaster.model.source.JavaClassSource;
 import org.jboss.logmanager.Level;
-import org.wildfly.apigen.invocation.ClientFactory;
-import org.wildfly.apigen.model.AddressTemplate;
+import org.wildfly.config.model.AddressTemplate;
 import org.wildfly.apigen.model.DefaultStatementContext;
 import org.wildfly.apigen.model.ResourceDescription;
 import org.wildfly.apigen.operations.ReadDescription;
