+++ generator/src/test/java/org/wildfly/apigen/test/AbstractTestCase.java
@@ -7,8 +7,8 @@ import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.wildfly.apigen.generator.Config;
 import org.wildfly.apigen.generator.ResourceMetaData;
-import org.wildfly.apigen.invocation.ClientFactory;
-import org.wildfly.apigen.model.AddressTemplate;
+import org.wildfly.apigen.generator.ClientFactory;
+import org.wildfly.config.model.AddressTemplate;
 import org.wildfly.apigen.model.DefaultStatementContext;
 import org.wildfly.apigen.model.ResourceDescription;
 import org.wildfly.apigen.operations.ReadDescription;
