+++ generator/src/test/java/org/wildfly/apigen/test/invocation/TypeAdapterTestCase.java
@@ -4,7 +4,7 @@ import org.jboss.dmr.ModelNode;
 import org.jboss.dmr.ModelType;
 import org.junit.Assert;
 import org.junit.Test;
-import org.wildfly.apigen.invocation.EntityAdapter;
+import org.wildfly.config.invocation.EntityAdapter;
 import org.wildfly.apigen.test.invocation.mail.subsystem.mailSession.MailSession;
 
 import java.util.Iterator;
