+++ runtime/src/main/java/org/wildfly/config/invocation/Types.java
@@ -1,18 +1,13 @@
-package org.wildfly.apigen.invocation;
+package org.wildfly.config.invocation;
 
 import com.google.common.base.CaseFormat;
 import org.jboss.dmr.ModelNode;
 import org.jboss.dmr.ModelType;
-import org.wildfly.apigen.generator.Keywords;
-import org.wildfly.apigen.generator.ResourceMetaData;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.logging.Logger;
-
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.TYPE;
 
 /**
  * @author Heiko Braun
@@ -142,24 +137,6 @@ public class Types<T> {
         }
 
         return result;
-    }
-
-    public final static String javaClassName(ResourceMetaData metaData) {
-        String name = null;
-        if(metaData.getDescription().isSingleton())
-        {
-            name = CaseFormat.UPPER_UNDERSCORE.to(CaseFormat.UPPER_CAMEL,  metaData.getDescription().getSingletonName().replace("-", "_"));
-        }
-        else
-        {
-            name = CaseFormat.UPPER_UNDERSCORE.to(CaseFormat.UPPER_CAMEL,  metaData.getAddress().getResourceType().replace("-", "_"));
-        }
-        return name;
-
-    }
-
-    public final static String javaAttributeName(String dmr) {
-        return CaseFormat.UPPER_UNDERSCORE.to(CaseFormat.LOWER_CAMEL, Keywords.escape(dmr.replace("-", "_")));
     }
 
     public static ModelType resolveModelType(Class<?> javaType) {
