+++ runtime/src/main/java/org/wildfly/config/model/AddressTemplate.java
@@ -19,7 +19,7 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.wildfly.apigen.model;
+package org.wildfly.config.model;
 
 
 import com.google.common.base.Joiner;
