+++ runtime/src/main/java/org/wildfly/config/model/ResourceAddress.java
@@ -19,50 +19,28 @@
  * Software Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
-package org.wildfly.apigen.model;
+package org.wildfly.config.model;
 
 import org.jboss.dmr.ModelNode;
 
-import static org.jboss.as.controller.descriptions.ModelDescriptionConstants.NAME;
-
 /**
+ * Represents a fully qualified DMR address ready to be put into a DMR operation.
  * @author Harald Pehl
  */
-public abstract class NamedNode extends ModelNode {
+public class ResourceAddress extends ModelNode {
 
-    private final String name;
+    public static final ResourceAddress ROOT = new ResourceAddress();
 
-    public NamedNode(final ModelNode node) {
-        set(node);
-        ModelNode name = get(NAME);
-        this.name = name.isDefined() ? name.asString() : undefinedName();
+    public ResourceAddress() {
+        super();
     }
 
-    protected String undefinedName() {
-        String undefined = "undefined_" + getClass().getSimpleName() + "_" + System.currentTimeMillis();
-        return undefined.toLowerCase();
+    public ResourceAddress(ModelNode address) {
+        set(address);
     }
 
-    public String getName() {
-        return name;
-    }
-
-    @Override
-    public boolean equals(final Object o) {
-        if (this == o) { return true; }
-        if (!(o instanceof NamedNode)) { return false; }
-        if (!super.equals(o)) { return false; }
-
-        NamedNode namedNode = (NamedNode) o;
-
-        return name.equals(namedNode.name);
-
-    }
-
-    @Override
-    public int hashCode() {
-        int result = super.hashCode();
-        result = 31 * result + name.hashCode();
-        return result;
+    public ResourceAddress add(final String propertyName, final String propertyValue) {
+        add().set(propertyName, propertyValue);
+        return this;
     }
 }
